requirements.txt
albumentations==1.1.0
av==9.2.0
h5py==2.10.0
imgaug==0.2.9
ipython==8.5.0
keras==2.10.0
keras_nightly==2.5.0.dev2021032900
keras_tuner==1.1.3
matplotlib==3.5.2
mrcnn==0.2
numba==0.55.2
numpy==1.22.3
opencv_python_headless==4.5.5.64
pandas==1.4.2
Pillow==9.2.0
scikit_image==0.19.2
scikit_learn==1.1.2
scipy==1.4.1
segmentation_models==1.0.1
segmentation_models_pytorch==0.2.1
skimage==0.0
tensorflow==2.10.0
tensorflow_gpu==2.9.1
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.0
